Feat: useStakesByFilterType hook
